rpmbuild-bot2: Create source subdir for each spec.
Previously, the in-place subdir along the spec file (usually, versioned) was used to override _sourcedir for rpmbild. This eliminated the need to copy auxiliary files from there to _sourcedir.
However, it exposed another problem: the ZIP file our scm_source RPM macros create from a software repository for the spec wolud end up there as well, "polluting" the auxiliary directory
with files that are not to be committed. Now a separate subdir is created in _sourcedir for each spec where auxiliary files are copied to (and were the ZIP is created) - just like the
old bot uded to. The difference is that there is a separate subdir per each spec so it's cleaner and no file name collisions are now possible for auxiliary files.