rpmbuild-bot2: Always git pull --no-rebase in upload.
This is essential as git pull may be set up to do rebase by default and this will always fail with a modified and uncommitted spec.
Also, pull is now done before add to ensure that if pull fails, the spec file will still be modified instead of staged on the
next run (to simplify diff show logic).