rpmbuild-bot2: Replace backslashes to forward slashes in common env. vars.
This is to avoid passing backslashes around because it's pretty difficult
to properly escape them in some cases and failures to do so may lead to
unexpected results.

One case is HOME which is normally set to something like X:\Users\user in CONFIG.SYS.
In RPMBUILD it becomes X:\Users\user/RPMBUILD/blah. Some of these paths may be passed
around without proper escaping or quoting. One case is -ffile-prefix-map GCC option
which stops working with backslashes because they are eaten together with the next
character in some cases.